% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(source, path)
}
\arguments{
\item{source}{A string specifying the source of the data. Can be "file" or "url".}

\item{path}{A string specifying the path to the data. For "file", it should be a local file path.
For "url", it should be a web URL.}
}
\value{
A summary of the data.
}
\description{
This function summarizes the input data from various sources.
}
